ExternTTS-fork: add personal fork as submodule
Added a personal fork of the previously used TTS
submodule, as it allows for immediate pushes.
The submodule follows the previous structure, so
it includes the same WaveRNN submodule as described
in the previous commit.

Updated the text2speech driver code in order to
account for the newly introduced TTS module (coming
from the `extern_tts_fork` submodule).

Signed-off-by: Gabriel Bercaru <gabriel.bercaru10@gmail.com>